Creates "ossec-register" common workflow
In order to support cross-host variable delegation, we'll need the play
to target *both* the Application and Monitor servers. We can then
dynamically open ports on the firewall as necessary in order to support
registration.

Reuses @msheiny's `iptables` Ansible module implementation, which is
remarkably clean. Love the state=absent functionality.

Rather than use `delegate_to`, we target both hosts and use the boolean
vars `ossec_is_server` and `ossec_is_client` (both defaulting to False)
in order to determine which host the task should execute on. For some
tasks, we want both hosts to execute, thus the combined play target.

Uses a rather gnarly hostvars-based retrieval to map a registered var
result across both hosts, ensuring the same value is accessible to both
play hosts, to coordinate the firewall rule management. Left an in-line
comment to guide future maintainers.

Removes most tags because we can better handle tasks at the import task
level now, which was difficult to do before because the logic was so
spaghetti.